+++ Demo/Demo/SwiftUIComponents/SwiftUIPaymentButtonDemo.swift
@@ -25,8 +25,7 @@ struct SwiftUIPaymentButtonDemo: View {
     private var labels = PayPalButton.Label.allCasesAsString()
     @State private var selectedLabel = PayPalButton.Label.allCases[0]
 
-    @ViewBuilder
-    var body: some View {
+    @ViewBuilder var body: some View {
         ZStack {
             VStack {
                 Picker("Funding Source", selection: $fundingIndex) {