+++ UnitTests/TestShared/MockAPIClient.swift
@@ -19,7 +19,6 @@ class MockAPIClient: APIClient {
         if let cannedFetchError {
             throw cannedFetchError
         }
-        // swiftlint:disable force_unwrapping
         let cannedData = cannedJSONResponse!.data(using: String.Encoding.utf8)!
         return try APIClientDecoder().decode(T.self, from: cannedData)
     }